+++ CHANGELOG.md
@@ -5,6 +5,8 @@ PowerModels.jl Change Log
 - Removed multi-network filter option from objective functions (breaking)
 - Removed option to run multi-network data in single-network models (breaking)
 - Improved robustness of matpower data parsing and transformation
+- Added support for three-winding transformers in PTI files
+- Added support for unit conversion in transformer entries of PTI files
 
 ### v0.6.0
 - Dropped support for Julia v0.5 (breaking)